SharedPreference 관련 로직 처리 코드 Repository로 이동
