+++ class14/onxmoreplz/HW2_Project/app/src/main/java/com/example/hw2_project/data/local/MovieLocalDataSourceImpl.kt
@@ -1,13 +1,30 @@
 package com.example.hw2_project.data.local
 
 import com.example.hw2_project.App
+import com.google.gson.Gson
+import org.json.JSONArray
 
 class MovieLocalDataSourceImpl : MovieLocalDataSource {
     override fun saveQuery(query: String) {
-        App.prefs.saveQuery(query)
+        val queryList = getSavedQuery().toMutableList()
+        queryList.add(query)
+
+        if (queryList.size > 5) {
+            queryList.removeAt(0)
+        }
+        App.prefs.saveQueryList(queryList)
     }
 
     override fun getSavedQuery(): List<String> {
-        return App.prefs.getSavedQueryList()
+        val json = App.prefs.getSavedQueryListTest()
+        val queryList = mutableListOf<String>()
+
+        json?.let {
+            val jsonArray = JSONArray(it)
+            for (i in 0 until jsonArray.length()) {
+                queryList.add(jsonArray[i].toString())
+            }
+        }
+        return queryList
     }
 }
